Create ClubList s a separate file
